ASK - Lab 1

+++ 2017-2018/ASK/8/zad1/zad1.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int clz(long); /* count leading zeros */
+long clz(long); /* count leading zeros */
 
 int main(int argc, char **argv) {
   if (argc < 2)
@@ -9,7 +9,7 @@ int main(int argc, char **argv) {
 
   long x = strtol(argv[1], NULL, 10);
 
-  printf("leading zeros of %ld: %d\n", x, clz(x));
+  printf("leading zeros of %ld: %ld\n", x, clz(x));
 
   return EXIT_SUCCESS;
 }
